# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AuthInfo:
    properties:
      created_at:
        type: string
      custom_pass:
        type: string
      edited:
        type: string
      email_sso:
        type: string
      id_level:
        type: integer
      jenis:
        type: string
      level_akun:
        type: integer
      level_kode:
        type: string
      niu:
        type: string
      updated_at:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  models.ListDetailResponse:
    properties:
      datas: {}
      messagae:
        type: string
    type: object
  models.ListResponse:
    properties:
      datas: {}
      pagination:
        $ref: '#/definitions/models.PaginationResponse'
      status:
        example: success
        type: string
    type: object
...
  models.PaginationResponse:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      pages:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  models.ProfileInfo:
    properties:
      email:
        type: string
      full_name:
        type: string
      key:
        type: string
      nidn:
        type: string
      nip:
        type: string
      niu:
        type: string
      user_type:
        type: string
    type: object
...
  models.RealisasiUnit:
    properties:
      _id:
        type: integer
      id:
        type: string
      kode_unit:
        type: string
      nama_unit:
        type: string
      pagu_pnbp:
        type: string
      pagu_rm:
        type: string
      pagu_rm_boptn:
        type: string
      realisasi_pnbp:
        type: string
      realisasi_rm:
        type: string
      realisasi_rm_boptn:
        type: string
      tahun_anggaran:
        type: string
    type: object
  models.RekapPMB:
    properties:
      _id:
        type: integer
      adikpapua:
        $ref: '#/definitions/models.StatPMB'
      internasional:
        $ref: '#/definitions/models.StatPMB'
      jumlah:
        $ref: '#/definitions/models.StatPMB'
      kode:
        type: string
      nama_prodi:
        type: string
      pasca:
        $ref: '#/definitions/models.StatPMB'
      profesi:
        $ref: '#/definitions/models.StatPMB'
      smbjm_cbt:
        $ref: '#/definitions/models.StatPMB'
      smbjm_raport:
        $ref: '#/definitions/models.StatPMB'
      smbjm_talent:
        $ref: '#/definitions/models.StatPMB'
      smbjm_utbk:
        $ref: '#/definitions/models.StatPMB'
      snbp:
        $ref: '#/definitions/models.StatPMB'
      snbt:
        $ref: '#/definitions/models.StatPMB'
      tahun:
        type: integer
      unit:
        $ref: '#/definitions/models.Unit'
    type: object
  models.StatPMB:
    properties:
      daftar:
        type: string
      lulus:
        type: string
      peminat:
        type: string
    type: object
...
  models.Unit:
    properties:
      fakultas:
        type: string
      fkt_kode:
        type: string
      jrs_kode:
        type: string
      jurusan:
        type: string
      prd_kode:
        type: string
      prodi:
        type: string
      uk_kode:
        type: string
    type: object
...
  models.UnitInfo:
    properties:
      fakultas:
        type: string
      fkt_kode:
        type: string
      jrs_kode:
        type: string
      jurusan:
        type: string
      prd_kode:
        type: string
      prodi:
        type: string
      uk_group:
        type: integer
      uk_id:
        type: integer
      uk_kode:
        type: string
      uk_nama:
        type: string
    type: object
...
  models.UserAuth:
    properties:
      _id:
        type: string
      auth_info:
        $ref: '#/definitions/models.AuthInfo'
      profile:
        $ref: '#/definitions/models.ProfileInfo'
      unit:
        $ref: '#/definitions/models.UnitInfo'
    type: object
...
host: localhost:8080
info:
  contact:
    email: anggadek857@gmail.com
    name: Dwi Angga
    url: https://github.com/anggananda
  description: |-
    This repository contains the **RESTful API** for the **Executive Information System (EIS) Undiksha**, built with **Golang** using the **Gin framework**.
    The API consumes data from **MongoDB** (populated via ETL from SQL in a separate repository) and uses **Redis** for caching.
    The system is integrated with **Undiksha Single Sign-On (SSO)** and implements **JWT Authentication** for secure access.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/anggananda
  title: Executive Information System RESTful API
  version: "1.0"
paths:
...
  /realisasi-unit:
    get:
      consumes:
      - application/json
      description: Mendapatkan data ralisasi unit berdasarkan filter dengan pagination
      parameters:
      - default: 1
        description: 'Nomor halaman (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Jumlah data per halaman (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Pencarian global
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ListResponse'
            - properties:
                datas:
                  items:
                    $ref: '#/definitions/models.RealisasiUnit'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get filtered ralisasi unit
      tags:
      - Realisasi Unit
  /realisasi-unit/export-csv:
    get:
      consumes:
      - application/json
      description: Mengekspor daftar realisasi unit yang telah difilter ke dalam file
        CSV.
      parameters:
      - description: Maksimal data yang akan diekspor
        in: query
        name: limit
        type: integer
      - description: 'Filter berdasarkan Tahun Ajaran (default: tahun sekarang)'
        in: query
        name: tahun
        type: string
      - description: Pencarian bebas
        in: query
        name: search
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File CSV berhasil diunduh
          schema:
            type: file
        "500":
          description: Kesalahan pada server saat pengambilan data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export Realisasi Unit ke CSV
      tags:
      - Realisasi Unit
  /rekap-pmb:
    get:
      consumes:
      - application/json
      description: Mendapatkan data rekap pmb berdasarkan filter dengan pagination
      parameters:
      - default: 1
        description: 'Nomor halaman (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Jumlah data per halaman (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Filter berdasarkan kode fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan kode jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan kode program studi
        in: query
        name: kodeProdi
        type: string
      - description: Filter berdasarkan tahun akademik
        in: query
        name: tahun
        type: string
      - description: Pencarian global
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ListResponse'
            - properties:
                datas:
                  items:
                    $ref: '#/definitions/models.RekapPMB'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get filtered rekap pmb
      tags:
      - Rekap PMB
  /rekap-pmb/export-csv:
    get:
      consumes:
      - application/json
      description: Mengekspor daftar rekap pmb yang telah difilter ke dalam file CSV.
      parameters:
      - description: Maksimal data yang akan diekspor
        in: query
        name: limit
        type: integer
      - description: Filter berdasarkan Kode Fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan Kode Jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan Kode Program Studi
        in: query
        name: kodeProdi
        type: string
      - description: 'Filter berdasarkan Tahun Ajaran (default: tahun sekarang)'
        in: query
        name: tahun
        type: string
      - description: Pencarian bebas
        in: query
        name: search
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File CSV berhasil diunduh
          schema:
            type: file
        "500":
          description: Kesalahan pada server saat pengambilan data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export Rekap PMB ke CSV
      tags:
      - Rekap PMB
...
  /user/details:
    get:
      consumes:
      - application/json
      description: Mendapatkan data profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ListDetailResponse'
            - properties:
                datas:
                  $ref: '#/definitions/models.UserAuth'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get data profile
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"